Refactor search tool architecture and introduce base classes
This commit removes the existing `search.py` file and introduces a new `_base.py` file that defines foundational classes and schemas for search tools. The new structure includes `SearchTool`, `SearchToolInputSchema`, `SearchToolOutputSchema`, and `SearchToolConfig`, which provide a standardized approach for implementing various search tools. Additionally, a new `QueryFocusStrategy` enum is added to facilitate query adaptation strategies. This refactor aims to enhance code organization and reusability across different search implementations.
